0.1 - raw data exploitation

+++ src/DLCompare/ApAppBundle/Resources/public/js/champion_list.js
@@ -16,7 +16,7 @@ apapp.champion_list.prototype.init = function()
 
 	$('ul#item-list.expanded li .long ol li strong').each(function()
 	{
-		var color = (parseInt($(this).html()) >= 0) ? "green" : "red";
+		var color = (parseFloat($(this).html()) >= 0) ? "green" : "red";
 		$(this).css('color', color);
 	});
 };
